Refactor with message update
