+++ HandleMessageViewUpdateFailures/handler.ts
@@ -1,28 +1,22 @@
 ﻿import { Context } from "@azure/functions";
-import {
-  MessageView,
-  MessageViewModel
-} from "@pagopa/io-functions-commons/dist/src/models/message_view";
-import { constVoid, flow, identity, pipe } from "fp-ts/lib/function";
+import { MessageViewModel } from "@pagopa/io-functions-commons/dist/src/models/message_view";
+import { constVoid, flow, pipe } from "fp-ts/lib/function";
 import * as TE from "fp-ts/lib/TaskEither";
-import * as E from "fp-ts/lib/Either";
 import * as t from "io-ts";
-import {
-  CosmosErrorResponse,
-  CosmosErrors
-} from "@pagopa/io-functions-commons/dist/src/utils/cosmosdb_model";
 import { MessageModel } from "@pagopa/io-functions-commons/dist/src/models/message";
 import { BlobService } from "azure-storage";
-import { RetrievedMessageStatusWithFiscalCode } from "../UpdateCosmosMessageView/handler";
 import { TelemetryClient, trackException } from "../utils/appinsights";
 import { errorsToError } from "../utils/conversions";
 import {
   Failure,
   PermanentFailure,
   toPermanentFailure,
-  toTransientFailure,
   TransientFailure
 } from "../utils/errors";
+import {
+  handleStatusChange,
+  RetrievedMessageStatusWithFiscalCode
+} from "../utils/message_view";
 
 const RetriableHandleMessageViewFailureInput = t.interface({
   body: RetrievedMessageStatusWithFiscalCode,
@@ -47,47 +41,6 @@ export type HandleMessageViewFailureInput = t.TypeOf<
   typeof HandleMessageViewFailureInput
 >;
 
-type CosmosErrorResponseType = ReturnType<typeof CosmosErrorResponse>;
-
-const isCosmosErrorPreconditionResponse = (
-  err: CosmosErrors
-): err is CosmosErrorResponseType =>
-  err.kind === "COSMOS_ERROR_RESPONSE" && err.error.code === 412;
-
-const isCosmosErrorNotFoundResponse = (
-  err: CosmosErrors
-): err is CosmosErrorResponseType =>
-  err.kind === "COSMOS_ERROR_RESPONSE" && err.error.code === 404;
-
-const wrapErrorToTransientFailure = (err: unknown): Failure =>
-  pipe(err, E.toError, toTransientFailure);
-
-const patchViewWithVersionCondition = (
-  messageViewModel: MessageViewModel,
-  messageStatus: RetrievedMessageStatusWithFiscalCode
-): TE.TaskEither<CosmosErrors, void> =>
-  pipe(
-    messageViewModel.patch(
-      [messageStatus.messageId, messageStatus.fiscalCode],
-      {
-        status: {
-          archived: messageStatus.isArchived,
-          processing: messageStatus.status,
-          read: messageStatus.isRead
-        },
-        version: messageStatus.version
-      },
-      `FROM c WHERE c.version < ${messageStatus.version}`
-    ),
-    TE.orElseW(
-      flow(
-        TE.fromPredicate(isCosmosErrorPreconditionResponse, identity),
-        TE.map(constVoid)
-      )
-    ),
-    TE.map(constVoid)
-  );
-
 export const HandleMessageViewUpdateFailureHandler = (
   context: Context,
   message: unknown,
@@ -111,95 +64,7 @@ export const HandleMessageViewUpdateFailureHandler = (
       )
     ),
     TE.map(retriableFailure => retriableFailure.body),
-    TE.chain(messageStatus =>
-      pipe(
-        patchViewWithVersionCondition(messageViewModel, messageStatus),
-        TE.orElseW(
-          flow(
-            TE.fromPredicate(
-              isCosmosErrorNotFoundResponse,
-              wrapErrorToTransientFailure
-            ),
-            // find and enrich message
-            TE.chain(() =>
-              pipe(
-                messageModel.find([
-                  messageStatus.messageId,
-                  messageStatus.fiscalCode
-                ]),
-                TE.mapLeft(wrapErrorToTransientFailure)
-              )
-            ),
-            TE.chain(
-              TE.fromOption(() =>
-                toPermanentFailure(
-                  Error(
-                    `Message metadata not found for ${messageStatus.messageId}`
-                  )
-                )
-              )
-            ),
-            TE.chain(messageWithoutContent =>
-              pipe(
-                messageModel.getContentFromBlob(
-                  blobService,
-                  messageWithoutContent.id
-                ),
-                TE.mapLeft(wrapErrorToTransientFailure),
-                TE.chainW(
-                  TE.fromOption(() =>
-                    toPermanentFailure(
-                      new Error(
-                        `Message body not found for ${messageWithoutContent.id}`
-                      )
-                    )
-                  )
-                ),
-                TE.map(content => ({ ...messageWithoutContent, content }))
-              )
-            ),
-            TE.map(messageWithContent => ({
-              components: {
-                attachments: {
-                  has:
-                    messageWithContent.content.legal_data?.has_attachment ??
-                    false
-                },
-                euCovidCert: {
-                  has: messageWithContent.content.eu_covid_cert !== null
-                },
-                legalData: {
-                  has: messageWithContent.content.legal_data != null
-                },
-                payment: {
-                  has: messageWithContent.content.payment_data != null
-                }
-              },
-              createdAt: messageWithContent.createdAt,
-              fiscalCode: messageWithContent.fiscalCode,
-              id: messageWithContent.id,
-              messageTitle: messageWithContent.content.subject,
-              senderServiceId: messageWithContent.senderServiceId,
-              status: {
-                archived: messageStatus.isArchived,
-                processing: messageStatus.status,
-                read: messageStatus.isRead
-              },
-              version: messageStatus.version
-            })),
-            // create message_view document
-            TE.chainEitherKW(
-              flow(
-                MessageView.decode,
-                E.mapLeft(flow(errorsToError, toPermanentFailure))
-              )
-            ),
-            TE.chainW(messageView => messageViewModel.create(messageView)),
-            TE.mapLeft(wrapErrorToTransientFailure)
-          )
-        )
-      )
-    ),
+    TE.chain(handleStatusChange(messageViewModel, messageModel, blobService)),
     TE.mapLeft(err => {
       const isTransient = TransientFailure.is(err);
       const error = isTransient
